Adding a few things to the build.jar
Lines 21 to 23: 
I noticed you didn't have the version following Semver, I highly recommend it, group and base not were not set either so I set both for you. 

Lines 18 to 19, 129 to 144: 
I have also added a json minify script for you, that will compress the various jsons in the assets and data folder, this helps decrease the overall size of the jar file. 

Lines 8, 16, 114-118, 146 - 165:
I have also added a new task you can run by doing `gradlew curseforge` this will build a jar for you and then upload it to the CF page for you. I have taken the time and set the game version to 1.16.3, marked it as Forge, and requiring Geckolib for you. All that is left is for Joosh to create an apiKey (https://www.curseforge.com/account/api-tokens) and for you to create a file called `key.properties`, which you want to mark as never upload to Github, which will look like so 
```
curseKey=xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
```
in which you put the token in place of the X's. Then simply just run the task and it will upload and set all the fields for you and put it into queue to be approved.